backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
aiosqlite==0.19.0
alembic==1.13.1

# Pydantic
pydantic[email]==2.6.1
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# AI
anthropic>=0.40.0

# Redis
redis==5.0.1

# S3/MinIO
boto3==1.34.34

# Utilities
python-dotenv==1.0.1
httpx==0.26.0

# WebSocket
websockets==12.0

# Image processing
Pillow==10.2.0

# PDF generation
reportlab==4.0.8

# Email
aiosmtplib==3.0.1
jinja2==3.1.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
